Allow admins to login as an exhibitor to make changes for them (#242)
* Update dependencies

* Allow admins to change exhibitor settings

* Make migration work with data in the database

* Localize

+++ src/server/api/routers/account.ts
@@ -72,12 +72,20 @@ export const accountRouter = createTRPCRouter({
         return { error: "invalidConfirmationCode" as const };
       }
 
+      const user = await ctx.prisma.user.findUnique({
+        where: { id: loginCode.userId },
+        select: { exhibitorId: true },
+      });
+      if (!user) {
+        return { error: "invalidConfirmationCode" as const };
+      }
+
       const [_, session] = await ctx.prisma.$transaction([
         ctx.prisma.session.deleteMany({
           where: { userId: loginCode.userId },
         }),
         ctx.prisma.session.create({
-          data: { userId: loginCode.userId },
+          data: { userId: loginCode.userId, exhibitorId: user.exhibitorId },
         }),
       ]);
